Refactor utils into a common folder
