+++ src/test_workflow/integ_test/service_opensearch_dashboards.py
@@ -18,7 +18,7 @@ from test_workflow.dependency_installer import DependencyInstaller
 from test_workflow.integ_test.distribution import Distribution
 from test_workflow.integ_test.distributions import Distributions
 from test_workflow.integ_test.service import Service
-from test_workflow.integ_test.utils import get_password
+from utils.version_utils import get_password
 
 
 class ServiceOpenSearchDashboards(Service):
